✨ 友達確認ページにて、友達の追加、削除を実装

+++ lib/src/components/friend_list_tile.dart
@@ -1,4 +1,6 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:sit_in_the_cafeteria/src/features/profile/pages/friend_list_notifier.dart';
@@ -21,19 +23,80 @@ class FriendListTile extends ConsumerWidget {
           SlidableAction(
             backgroundColor: Colors.red,
             onPressed: (_) {
-              // 削除処理
-              friendListNotifier.deleteFriend(friendID);
+              // iosダイアログの表示
+              showCupertinoDialog(
+                context: context,
+                builder: (context) => CupertinoAlertDialog(
+                  title: const Text('友だち削除'),
+                  content: const Text('友だちを削除しますか?'),
+                  actions: [
+                    TextButton(
+                      onPressed: () {
+                        Navigator.pop(context);
+                      },
+                      child: const Text('キャンセル'),
+                    ),
+                    TextButton(
+                      onPressed: () {
+                        // 削除処理
+                        friendListNotifier.deleteFriend(friendID);
+
+                        Navigator.pop(context);
+                      },
+                      child: const Text(
+                        style: TextStyle(color: Colors.red),
+                        '削除',
+                      ),
+                    ),
+                  ],
+                ),
+              );
             },
             icon: Icons.delete,
             label: '削除',
           ),
         ],
       ),
-      child: Container(
-        color: Colors.white,
-        child: ListTile(
-          title: Text(friendName),
-          subtitle: Text(friendID),
+      child: InkWell(
+        onTap: () {
+          // 友達のIDをコピー
+          Clipboard.setData(ClipboardData(text: friendID));
+
+          // スナックバー表示
+          ScaffoldMessenger.of(context)
+            ..removeCurrentSnackBar()
+            ..showSnackBar(
+              SnackBar(
+                content: Text('$friendNameのユーザーIDをコピーしました'),
+              ),
+            );
+        },
+        child: Container(
+          decoration: const BoxDecoration(
+            border: Border(
+              bottom: BorderSide(
+                color: Colors.grey,
+                width: 0.5,
+              ),
+            ),
+          ),
+          height: 60,
+          width: double.infinity,
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                style: TextStyle(
+                  color: Theme.of(context).colorScheme.secondary,
+                  fontSize: 18,
+                  fontWeight: FontWeight.w500,
+                ),
+                friendName,
+              ),
+              Text('ID  $friendID'),
+            ],
+          ),
         ),
       ),
     );
